+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A bunch of SDK tests are skipped due to regressions following the decoupled builder merge ([#306](https://github.com/Substra/substra-tests/pull/306))
 - BREAKING: replace `todo_count` and `waiting_count` by the new counts following the new statuses in the backend ([#319](https://github.com/Substra/substra-tests/pull/319))
 
+### Added
+
+- Test case for function build on another backend than the task owner ([#318](https://github.com/Substra/substra-tests/pull/318))
 ## [0.47.0] - 2023-10-18
 
 ### Added